[NES] add support for the Zapper
The Zapper is much jankier than the Super Scope, as you'd expect from a previous generation of hardware. It only exposes 2 bits to software: whether the trigger is currently half-pulled and whether the light sensor recently detected a bright light (roughly within the last 20-25 PPU lines).

Games detect the Zapper position by briefly turning most of the screen black, placing bright-colored boxes over possible targets, and polling to see when (or if) the Zapper detects a light while the PPU is rendering.