[32X] fixes for Brutal Unleashed: Above the Claw, a bad game that also seems to be badly programmed
* run the two SH-2s more in lockstep, which is necessary because of how it uses the communication ports
* stub out the SH7604 SCI registers; this game writes to them but doesn't seem to use the SCI
* various minor performance improvements

The communication port issue: Whenever the master SH-2 writes to one of the communication ports, the slave SH-2 will see it and then write to one of the communication ports twice in quick succession. The master SH-2 needs to read the first write or it will desync. This immediately breaks background scrolling and it will cause the game to freeze at the end of the fight.