[GEN/SCD/32X] emulate memory refresh delay, plus general timing improvements for SCD/32X by emulating bus arbiter delay
this is only approximate emulation because exact emulation is difficult and because the precise timings are not publicly known. the 68000 is now stalled for 2 out of every 128 mclk cycles, which is roughly the delay that it would get when executing out of ROM

Sega CD gets a less frequent delay (2 out of every 172 mclk cycles) because that's what seems to make mcd-verificator tests pass. it's not currently clear to me why this is or if this is even correct