[SNES] fix implementation of Mode 7 scrolled center point clipping (#426)
Clipping to i10 is not correct because that breaks various screens in Super Metroid. The correct behavior is to combine the magnitude from the lowest 10 bits and the sign from the highest bit. This fixes Super Metroid and doesn't break Kaite Tsukutte Asoberu Dezaemon or Tiny Toon Adventures: Wacky Sports Challenge